lab5 : 可以从内核线程跳到用户进程，但用户进程间切换还有问题

+++ labcodes_answer/lab5/kern/init/init.c
@@ -59,7 +59,7 @@ grade_backtrace2(int arg0, int arg1, int arg2, int arg3) {
 
 void __attribute__((noinline))
 grade_backtrace1(int arg0, int arg1) {
-    grade_backtrace2(arg0, (int)&arg0, arg1, (int)&arg1);
+    grade_backtrace2(arg0, (int64_t)&arg0, arg1, (int64_t)&arg1);
 }
 
 void __attribute__((noinline))
@@ -69,7 +69,7 @@ grade_backtrace0(int arg0, int arg1, int arg2) {
 
 void
 grade_backtrace(void) {
-    grade_backtrace0(0, (int)kern_init, 0xffff0000);
+    grade_backtrace0(0, (int64_t)kern_init, 0xffff0000);
 }
 
 static void
